Fix very nasty bug
A list comprehension was filtering by {ok, X}, but the actual results
were just X, so the list comprehension skipped all the results. This led
to notifications silently failing to be sent if they relied on the
registration mechanism.

Since SCPF has only been used with non-registered notifications, this
wasn't an issue, but the bug went unnoticed until the test cases were
resuscitated.

This commit fixes the issue.