Add async_send/2,3; fix type and fun specs
Also fix dialyzer warning that arose because `start_timeout`
was being extracted from a list of proplists, which made dialyzer
quietly decide the type of that list of proplists was just a proplist,
causing a type warning when calling a function later that expected
a list of proplists. The fix is to extract `start_timeout` from each
proplist, rather than from the list of proolists.